Типы офера вынесены в директорию types
