+++ src/components/favorites-offer-card/favorites-offer-card.tsx
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import type { ServerOffer } from '../../mocks/offers';
+import type { ServerOffer } from '../../types/offer';
 
 import { AppRoute } from '../../const';
