+++ src/hooks/use-map.tsx
@@ -1,8 +1,6 @@
-import {useEffect, useState, useRef, MutableRefObject} from 'react';
+import { useEffect, useState, useRef, MutableRefObject } from 'react';
 import { Map as LeafletMap, TileLayer } from 'leaflet';
-import { ServerOffer } from '../mocks/offers';
-
-export type City = ServerOffer['city']
+import { City } from '../types/offer';
 
 function useMap(mapRef: MutableRefObject<HTMLElement | null>, city: City): LeafletMap | null {
 	const [map, setMap] = useState<LeafletMap | null>(null);
